Unpinned aenum dependency
ethanfurman/aenum#13 was fixed in version 3.1.1, so just skip versions 3.0.* and 3.1.0 and all should be well.

Also fixed a test that broke on the new aenum version: in a change of behaviour, the zero flag is no longer treated as being "in" a combination of flags (not even a combination that includes the zero flag). This wasn't a very idiomatic thing to expect anyway, so that's fine.

Honestly I'm not even sure what that test was meant to achieve. Maybe I was checking that the conformant property is defined? Or I copied-and-pasted the test for the renderable_type property, and unthinkingly substituted NONE since it's the default value according to the spec?